MCP/requirements.txt
mcp>=1.2.0
google-genai>=0.5.0
pydantic>=2.0.0
